requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
python-dotenv==1.0.0
google-generativeai==0.3.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
requests==2.31.0
aiofiles==23.2.1

